+++ gulpfiles/js/module/cookieconsent.js
@@ -35,8 +35,8 @@ function initCookieConsent() {
                 href: cookieLayer.href
             },
             onInitialise: function(status) {
-                var type = this.options.type;
-                var didConsent = this.hasConsented();
+                let type = this.options.type;
+                let didConsent = this.hasConsented();
                 
                 if (type === 'opt-in' && didConsent) {
                     // enable cookies
@@ -58,8 +58,8 @@ function initCookieConsent() {
                 // do something
             },
             onStatusChange: function(status, chosenBefore) {
-                var type = this.options.type;
-                var didConsent = this.hasConsented();
+                let type = this.options.type;
+                let didConsent = this.hasConsented();
                 
                 if (type === 'opt-in' && didConsent) {
                     // enable cookies
@@ -75,7 +75,7 @@ function initCookieConsent() {
                 }, 1000);
             },
             onRevokeChoice: function() {
-                var type = this.options.type;
+                let type = this.options.type;
                 
                 if (type === 'opt-in') {
                     // disable cookies