+++ gulpfiles/js/module/langswitch.js
@@ -6,7 +6,7 @@
  */
 (function($) {
     if (document.querySelector('html').classList.contains('locale-process-session')) {
-        for (var i = 0; i < document.querySelectorAll('a.langswitch').length; i++) {
+        for (let i = 0; i < document.querySelectorAll('a.langswitch').length; i++) {
             document.querySelectorAll('a.langswitch')[i].onclick = function(event) {
                 event.preventDefault();
                 setCookie('current_locale', this.getAttribute('data-locale'), 365); // eslint-disable-line no-undef